refactor(arweave client): rename to ChainApiClient to ArweaveCompositeClient
The name reflects the fact that this Arweave client reaches out to both
the trusted node and peers. In the future the peer and trusted node
logic may be extracted and composed in this class, but for now it's
still all combined.

Also, I removed the "Chain" from the name because I anticipate adding
data related functionality in the future (e.g. fetching chunks).

PE-1802