feat(graphql sqlite): add ability to query recent blocks PE-1678
This adds the ability to query recent blocks by running an additional
query for the new or stable blocks depending on the sort order if the
result set is less than the page size.

The queries are run independently both for ease of implementation (the
independent queries are simpler and easier to understand) and
performance (additional tables are queries only when needed).

This also fixes a minor issue when querying for a single block with a
non-existant ID.