feat(migrations sqlite): add custom umzug storage implementation PE-1950
Adds a custom umzug storage implementation that uses better-sqlite3 to
store migration status in 'core.db'. Technically I think we could use
the Sequelize storage implementation too, but that feels like a pretty
heavy dependency to pull in just to keep track of migration status. JSON
is an option too (I used that for initial testing), but I worry about
users accidentally nuking the JSON file with that approach.