pref(sqlite graphql): join tags with lots of values last
There are certain tags like Content-Type and App-Name where we know that
a given name value pair may have a lot of records (e.g Content-Type:
text/html). SQLite isn't very good at figuring this out at run time and
doing joins in the optimal order, so we help it out by telling it the
join order.

We'll likely do an inverse version of this too where we use the shanon
entropy of the tag value as a proxy to figure name value pairs with very
low frequency.