refactor(sqlite): pass path instead of DB into constructor
Doing this because the Sqlite DB object won't be able to cross the
thread boundary into the soon to be added worker threads. This
unfortunately slows down the tests slightly (still runs in 3s). I'll
look into truncating tables instead of recreating the DB each to see if
that improves performance at all.