test(block-importer): cleanup DB file after every test
Still actually need a bit more cleanup to remove -shm and -wal files.
That's trickier because the block importer doesn't close the DB after
each run. I'm going to punt on that for now since I plan to swich to a
single DB that gets truncated on each run anyway.