perf(sqlite): clear height on new_blocks when reseting for forks
Previously we used height from new_block_heights as the canonical height
for new_blocks. For performance reasons, we are moving away from that in
favor of using height on new_blocks (joins slow us down more than
updates will). In order to accomplish that we need to make sure we clear
heights on new_blocks > fork height after a fork occurs.