feat(chunks): extract data_root and chunk sha256 from tx_path and data_path PE-2284
We we could technically use the data_root passed in, but this let's us
avoid an extra base64 decode step. The hash could also be computed from
the data, but extracting it from data_path both reduces the compute
required to get the hash and allows us to extract hashes from the chunk
metadata alone.