refactor(chunks cache): remove unnecessary catch + rethrows PE-2284
Whoops, some copy-pasta here on my part. In general, it's not a good
idea to catch and rethrow. Use a finally instead if cleanup is needed.
In this case it was done for logging purposes, but the caller can always
log the error if they choose. However, they can't choose not to log it
if we log in the catch.