refactor(composite-client): remove unnecessary error logging PE-2284
Removing more catch + log + rethrow - we should let the caller handle
this if desired.

Note: there are a couple legitimate exceptions to this rule in the
composite-client, but they're in cases where we have to handle the error
or it will turned into an unhandled promise rejection. There are also a
couple exceptions in the store classes, but that may change. They're
that way now because we often use them as caches and we don't want cache
misses caused be errors to accidentally blow up callers.