refactor(routes data): use /raw/<txid> for l1 data testing PE-2476
If we use /<txid> for data testing then we need to pass requests we
can't find to arweave.net (or another gateway), this is cubmersome
because Envoy doesn't provide us with an easy way to fall back other
than load balancing which has the side effect of sending some of the
traffic to arweave.net even when not falling back. Since /raw isn't
actively used by applications yet, it's a better target for testing (we
don't need a fallback for it).