refactor(data): add 'verified' to contiguous data return type PE-2797
Verified is set to true if we have verified that the data returned
matches the expected TX signature or data root. In the case of data
constructued from contiguous chunks, if all the chunks validate then we
know this is the case. Note, that this implies all chunk sources must
verify the chunks they return.