feat: add a weak map based chunk cache PE-2809
This helps if you request the same chunk in rapid succession or if you
read both the metadata and data for the chunk and the requests fall
through to a node. Using a weak map means this should incur no extra
memory overhead since the cached chunks can be GCed at any time.