refactor(data): add manifest path resolution interface PE-2810
Adds a ManifestDataPathResolver interface to enable support for multiple
path resolver implementations (for testing and to support different DBs).
This is one of what I expect to be two manifest path resolution
interfaces. This one accepts data so it can parse on demand, the other
(if/when it's added) will take just an id and path and will look only in the DB.
The latter is only needed for optimization so that we can skip the data
request when the manifest is already indexed.