refactor(data): add interface method for resolving manifest path from index PE-2810
Originally I planned to make this two interfaces, but that feels like
speculative generality since the data route requires both.

The current streaming implementation just has a stub for the resolveFromIndex
method. After manifest DB indexing is implemented the sreaming path
resolver will be refactored (and renamed) to check the DB index too.

This change also introduces the concept of "complete" resolution. It's
possible for the DB index to be incomplete due to lazy indexing,
in-progress batch indexing, or parsing issues. In those cases we should
fall back to on-demand indexing. However, batch imports can keep track
of how many paths there are in a manifest as well as how many they've
imported. If they match, we know the index is complete for that manifest
and don't need to try parsing it on the fly.