feat(data): use data attributes (e.g. content type) from DB if they're available PE-2810
The DB data index (TX + data item metadata) has the authoritative content types,
sizes, etc. We should use them if we've already indexed the data in
question. If we have not, we can fall back to what the data source provides, but
that should be a last resort.

This also extracts repeated manifest handling code into a standalone
function.