refactor(data): add isManifest to data attribute PE-2810
This moves the content type comparison logic for data attributes into
the getDataAttributes method. That allows us to potentially use
something other than content type to make the determination about
whether a particular ID is a manifest. In the future we'll likely check
the manifest index in the DB too.