feat(data): add restrictive raw data headers PE-2810
This mirrors the behavior on arweave.net by adding restrictive cross origin
and content security policies for raw data. The raw interfaces is
intended for use by other services as rather than web apps.

Note: there is one restriction that's still not implemented - limiting
default-src to only the primary domain (not ArNS or manifests domains).
That requires an extra configuration option to tell the service about
that domain and we do not yet have a config option for that.