feat(data): add contiguous sequntial data source PE-2810
Tries contiguous data source one at a time until it finds one that
returns something. We currently instantiate this such that it tries a
trusted gateway first (for performance reasons). However, eventually, it
would be nice if we tried chunks first since chunks are the easiest data
source to verify.