feat(envoy): add optional GraphQL pass through support PE-2810
Add support for specifying a GRAPHQL_HOST and GRAPHQL_PORT in .env. If
they are set, GraphQL traffic will be passed through to the specified
target.

Also, adds the ability to override other environment variables via .env
(check docker-compose.yaml for the list).

Note: This required switching to ytt for YAML templating. Also, there is
some special case logic related to proxying GRAPHQL to arweave.net
specifically. Hopefully this can be simplified or removed in the future.