feat(arns): add ArNS resolution via trusted gateway
Resolves ArNS names by sending a HEAD request to a trusted gateway. The
X-ArNS-Resolved-Id and the X-ArNS-TTL headers on the response are used
to determine the TX or data item ID for the name and the max-age to set
in cache headers.

Resolution occurs via a middleware that intercepts subdomain requests
and passes them on to the data handler. The data handler detects the
resolved ID and uses it instead of the ID from the route parameters.

Name resolution in the middleware is handled by an implementation of a
NameResolver passed into the middleware when it is created. This enables
supporting multiple resolution strategies in the future.