test(end-to-end): add script to run all end-to-end tests PE-2855
Also renames 'functional' to 'end-to-end' - I think these could
technically be considered 'functional' tests by some definition, but
'end-to-end' seems a lot less ambiguous. The point of them is to
verify that all the real system components run and work together as
expected when run in a manner similar to how users run them.