refactor(arns): limit subdomains to 30 or fewer characters
This is a sanity check to ensure we don't somehow get a name that
overlaps with the base32 encoded TX ID URLs used for sandboxing. The
ArNS contract itself imposes a shorter 20 character limit.

The previous sanity check limit was < 43. That's confusing because TX IDs
are 43 characters and this limit is not directly related to that specific
length.