feat(moderation): add content moderation enforcement to data path PE-3235
Add DB tables to record blocked content by ID and hash and use those
tables in the data handlers to prevent serving blocked content.
The ability to store notes and sources for blocks is also supported.

Note: These is no API for adding blocked IDs and hashes yet. That will
be implemented in a subsequent commit.