feat(moderation): add admin endpoint to block by id or hash PE-3235
Adds an admin endpoint (/ar-io/admin/block-data) that accepts a JSON
body containing either an id or hash to block along along with optional
notes and a source name (e.g. a blocklist). In the future we may add
a worker process to reads from a remote list, but this is sufficient to
support basic content moderation requirements.