feat(bundles): add an ANS-104 unbundling worker
The worker unbundles layer 1 transactions with 'Bundle-Format: binary'
and 'Bundle-Version' starting with '2.'. A new TX matching interface is
used to implement this (this will be used for other TX matching tasks in
the future). Unbundling consists of parsing the bundle and creating
events with data item metadata. No actual indexing happens during
unbundling. Instead indexing will be performed by consumers of the data
item events.