feat(bundles data): add tables to relate parent/child data PE-3477
Adds two tables - one that relates data hashes to their parent hashes
and another that relates ids to their parent ids. A parent hash/id in
these tables means the data for the id (not the parent) is retrievable
from the parent by reading a given amount of data starting at a
particular offset in the parent. The hash table contains an offset but
no size since the contiguous data hash table already includes the size.
The id table includes both the offset and size since we don't record
size for IDs (we rely on the related hash data size and we can have ids
without hashes).