feat(bundles data): add ANS-104 data item data indexer PE-3477
This adds a new interface for saving data item id, parent id, offset,
and size and a worker that uses that interface to index data item data
in response to data item unbundle events. Note that the worker currently
responds directly to events. That's a shortcut to get the indexing logic
wired together more quickly. Future revisions will place a queue between
the event handler and the worker.

This revision also includes some type refactoring - TransactionLike was
renamed to MatchableItem, and TransactionFilter was renamed to
ItemFilter. This better reflects the purpose of the MatchableItem interface
(matching items - TXs + data items) and the ItemFilter filter interface.
I still dislike how general 'Item' is, but in context it seems relatively
clear. If I think of something better I may revise it further.