refactor(bundles): adjust NormalizedDataItem to match TX format PE-3477
The purpose of the NormalizedDataItem interface is to provide a common
interface for ANS-102 and ANS-104 data items as well as make them
resemble transactions as much as possible. As such, we should match the
naming conventions on TXs as closely as possible. At some point, I could
see us refactoring all of this to use zod and normalizing all the naming
(removing underscores, etc.), but I'm uncomfortable doing that without
some stronger type checking around the edges of the systems.