refactor(db): standardize index/db names PE-3477
- Top level chainDb is now just db
- ChainDatabase is now ChainIndex
- NestedDataIndexer is now NestedDataIndexWriter

The top level name change was because it's not just a chainDb anymore.
The interface name changes are for consitency and because something
can serve as an index without actually being a DB. And, finally, changing
Indexer to IndexWriter makes it clear that the interface is only used for
writes to the index.