refactor(ans-104): use a queue when indexing data item data PE-3600
This places a queue between the event handler and the ans-104 data
indexer. Instead of handling event immediately they are placed in a
queue and the ans-104 event handler processes them one at a time (per
consumer). Currently this serves primarily as a way to limit concurreny
since we use on an in memory queue without any back-pressure or load
based scaling. In the future we will support more flexible queueing
systems.