refactor(ans-104): use a queue when unbundling ans-104 bundles PE-3600
Similar to the earlier data item indexer change, this places a queue
between the event handler and the ans-104 unbundler. Instead of handling
event immediately they are placed in a queue and the ans-104 event
handler processes them one at a time (per consumer). Currently this
serves primarily as a way to limit concurreny since we use on an in
memory queue without any back-pressure or load based scaling. In the
future we will support more flexible queueing systems.