refactor(events): replace '*-saved' with '*-indexed' and use constants PE-3600
Renames '*-saved' to '*-indexed'. 'saved' was used originally because
it matches the methods on the DB/index class. However, as we're starting
to add events related to writing and indexing the data itself, 'saved'
is starting to feel ambigious. Also, as part of this renaming, all the
event strings have been moved into event.ts and assigned to constants.
This both reduces likelihood of typos and makes future refactoring
easier.