feat(ans-104): add configurable unbundling filters PE-3625
Adds two new environment variables for configuring which ANS-104
bundle transactions get unbundled and what data item data gets
indexed (see the README for details).

The environment variables contain JSON representing the filter. Filters
are described by a map containing a single key, the match operator, and
a value representing the arguments to those operators. Some match
operators like the 'and' and 'or' operator accept arguments that are
themselves filters. In those cases, the arguments are evaluated
recursively as filters. If a match operator takes no arguments 'true' is
used as the argument. The following filters are supported: always,
never, and, or, tags, attributes.

Example:

  {
    and: [
      {
        tags: [
          { name: "foo", value: "bar" }
        ]
      },
      {
        tags: [
          { name: "baz", valueStartsWith: "qux" }
        ]
      }
    ]
  }