refactor(sqlite): rename s/high cardinality/low selectivity/ PE-2395
'High cardinality' is confusing because it's not obvious if it's
referring the tag name + value pair cardinality or the TX ID cardinality
(former is low, latter is high). 'Low selectivity' is clearer. If you
are querying on the specified tags, you should expect a large result set
(for that part of the query). Thus it would be better to filter on other
criteria first.