refactor(system): export constants for DB interfaces PE-3601
As we start to support other DBs, we will likely have a variety of
different implementations of the index interfaces. To enable this, we
should export a different constant per DB interface. That will reduce
the need for future refactoring since each part of the system will
depend only on the constant that fulfills the interface it needs.

Note: there are still a couple more places this needs to happen, but this
is moving in the right direction.