feat(bundles): move unbundling to a worker thread PE-3601
Moves unbundling to a worker thread to avoid monopolizing the main
thread. Currently uses a temp file to pass the bundle into the worker
thread. In the future, bundles will be streamed to the worker using
shared buffers. Only 1 unbundling thread is currently support (also
likely to change in the future).