feat(bundles): improve data item height tracking + optimize stable flushing PE-3769
Clears the heights on data items > fork height when forks occur and
updates data items related to L1 TXs when L1 TX heights are set. Also
adds a height condition to the query for data items to flush to stable
to avoid unnecessary work when joining to L1 stable tables to retrieve
cannonical heights.

Note: further optimization may still be possible. It may be possible to
eliminate one of the joins by replacing it with a join to
stable_block_transactions if we add height to stable_block_transactions.
Though, it's unclear how much performance improvement that would yield.