feat(graphql): increase page size limit to 1000
Raise the GraphQL page size limit from 100 to 1000. There's very little
performance penalty for doing this. In fact, it's probably better for
performance overall if applications to issue fewer queries and return
larger result sets. Also, if we raise the limit to 1000 then it's easy
for application to return entire blocks at a time (blocks contain a max
of 1000 transactions).