feat(sqlite bundles): add ability to query stable data items PE-3769
Combines stable transactions and stable data items using a UNION in the
SQL query. Each subquery in the UNION has its own ORDER BY and LIMIT.
This allows the sub-selects to do most of the work before the union is
computed. This change also implements returning parent/bundleIn for data
items. However, filtering based on bundledIn and sorting data items by
ID are not functional yet and will be implemented in future commits.