feat(sqlite graphql): include data items in sorting and cursors PE-3769
Adds data items to GQL sorting and cursors. Data items are sorted by ID
after block height and block TX index. ID was chosen as opposed to
bundle offsets or indexes because we want duplicates of the same item
sorted consistently where possible. Also, bundle data item indexes are
potentially confusing when data item filtering is used.

In order to accomplish this, the cursor condition in the query was
changed from a simple numeric comparison to a set of comparisons against
the cursor components. An OR is required in the comparisons to avoid
comparing against irrelevant conditions (e.g. block TX index comparison
when height > cursor height). This clutters the WHERE conditions, but is
still fairly readable. Also it may perform better since it makes the
height comparison legible to SQLite.