feat(sqlite graphql): support querying "new" data items PE-3769
Adds support for querying data items that have not yet been flushed to
the stable (> 50 blocks old) tables. Note, there are still some edge
cases to work out with this and new data querying in general. In
particular, we don't currently support querying data that has not yet
been associated with a block or data is technically stable but was
indexed late (e.g. due to missing chunks) and has not yet been flushed.