feat(sqlite bundles): record all parent/child relationships for matching data items PE-3769
We don't want a data item with the same ID to appear multiple times in
GraphQL, so we only insert unique IDs into new_data_items. However, we'd
still like to have a record of all the bundles containing a particular
ID. This is important if a bundle is removed (due to content moderation)
or the parent association needs to be changed for any other unforeseen
reason.