refactor(bundles ans-104): push filtering down into worker PE-3769
This moves filtering down into the parser so that we can (in a future
commit) emit an event that indicates how many data items within each
bundle matched the filter. We want that in order to detect bundles that
failed to import successfully. There are a couple of side benefits of
this too - 1. it moves more work out of the main thread; 2. it reduces
the amount of messages that go back to the main thread.