feat(bundles): add bundle process tracking PE-3769
Records bundle records that include first and last timestamps for
queuing, skipping, unbundling, and indexing (note: indexing timestamp
column is present, but not yet set). Data items counts, both total and
matched by the index filter, are also recorded as well as the IDs of the
filters used to match both the bundle and the data items in it. These
can be used later to decide when to reprocess bundles.

Note: 'last_fully_indexed_at' is handled slightly differently from other
'last_*' timestamps. Most are not overwritten if they're already set but
'last_fully_indexed_at' is. It's assumed that if the bundle record is
being updated in some way it means the bundle is being reprocessed and
it's indexing status should be cleared unless it's explicitly set as
part of the update.