fix(bundles data): fix infinite recursion when parent data is missing PE-4054
The recursive case when getting parent data was incorrectly passing the
original ID instead of the parent ID. That lead to infinite recursion
since it was continually finding the same parent and then trying to
download it. This change corrects that and fixed what appeared to be an
issue with setting passing the size for nested bundles. The size should
always be the original size. It's only the offset that should be added
to during recursion.