feat(bundles): add a process to reindex bundles after a filter change PE-4115
Adds a process that resets bundle timestamps for bundles that were
processed with different filters than are currenly in use. Since the
process creates some DB load even if the filters are unchnaged, it is
only enabled when the FILTER_CHANGE_REPROCESS environment variable is
set to true. In the future we may optimize this further by keeping a log
of filter changes. That would enable more efficient queries based on
comparing timestamps (< filter change time) rather than filter IDs
(using an inequality).