refactor(bundles ans-104): use owner address from data item instead of rehashing
Prior to this change we were hashing the owner key to get the owner
address. This change uses the owner address from the data item instead.
These should always be the same value so rehashing is unnecessary.

Note: I ran a test comparing the values and on the sample of data items
I processed there were no differences.