perf(graphql): correct GQL query performance issue
When ANS-104 support was added, the order by statements were adjusted to
not specify the stables to use for order (column numbers were used
instead). When filtering by tags and querying by stable data, this
introduced a performance regression. In that case, it's important to
sort by the height in tag tables for the best performance. This change
reintroduces sorting by table alias and field instead of column number.