fix(bundles): use data item content type when available PE-4404
When bundles are indexed, the data items offsets are saved so that they
can be used later to read data item data from the appropriate place in
the contiguous L1 data. In order to respond with appropriate content
type on that data we need to retrieve the content type from the data
item index. Prior to this change we were incorrectly falling back to the
default content type in that situation.