fix(envoy): use 30 second cache control header for proxied requests PE-4552
When using a CDN we want to make sure that we do not use a long cache
duration for endpoints like `/current_block` and `/height` where the
response changes as each new block is produced.

This change adds a Cache-Control max-age of 30 seconds for all proxied
requests. In the future we will adjust this to be route specific
(responses for block and TX by ID can be cached longer).