feat(info observer): advertise the contract ID in use PE-4894
This allows gateways to indicate which contract they are using. Of
course, there is no way to verify that the gateway is indeed using that
contract for anything. But the contract state includes the hostname, so
if the contract state includes the gateway in question and the gateway
points to the contract, you can reasonably assume that the wallet holder
and the gateway operator desire to use that contract.

Note: this is not true on the observer side. The contract state does not
contain a hostname for the observer (it could be different from the
gateway host). Thus, the observer advertised contract ID should be
viewed as a convenience only. On-chain interactions are how observers
indicate which contract they are using.