revert(lmdb): disable compression and simplify parameters passed to LmdbKVStore
After some simple testing (starting at 800k and running for 750 blocks) there is no material difference in database filesize with `compression` enabled. Moving other parameters to class and only allowing `dbPath` to be provided simplifies the interface and leans on LMDB defaults, which should be fine for this initial implementation.