feat(block-store): implement `KvBlockStore` and use it with `lmdb` implementation
Note: there is currently a bug fetching block data by height. Something is going wrong with encoding the block hash against the height causing for prefetching to fail for certain blocks (e.g. 800079). Original assumption it was an encoding issue converting strings to buffers, but that did not solve the problem. Continuing to investigate but pushing this commit to get eyes from others.