fix(lmdb): guarantee returning a buffer from lmdb
While investigating why some blocks failed to unpack/decode correctly from the KvBlockStore we found what appears to be a bug in the lmdb library. Under the hood, Buffers can be converted to Uint8Array and it is not clear when that happens (or why). This causes our decoding to fail and all the base64 encoded strings to return the buffer array as a string, rather than their decoded value. The behavior seems to be consistent for certain blocks, (e.g. `800076`, `800079`, etc.) but  undetermined if its related to the size of the block object being stored. As a simple workaround we guarantee the return of a `Buffer` object by using `Buffer.from()`. `value.buffer` uses the same memory space as the existing Uint8Array, avoiding duplication of the buffer data.