tests(routes data): add example data handler test PE-4908
Adds a minimal test that checks ID blocking behavior with the goal of
serving as an example for range request PR tests. For now it uses stubs.
That was done for expedience only. In the future it will be updated to
use the the real implementations of the data handler dependencies.