feat(ans-104): add support for parallel unbundling PE-4903
Adds worker threads to the Ans104Parser implemented in a similar fashion
to the SQLite workers. There is a fixed size worker pool. Each worker
takes work as it becomes available, processes it, and then repeats.
Callers block until work is complete. The number of workers defaults to
2 but is configurable via the ANS104_UNBUNDLE_WORKERS environment
variable. Along with this change, The number of data item indexing
workers is increased from 1 to 2 to help ensure that communication round
tip latency between the main thread the DB threads is not the bottleneck
for data item indexing.