feat(validation): use regex to sanity check block hashes
The format sanity check functions are supposed to be lightweight and
performant and should therefore not exhaustively check the format of
transactions and blocks. However, we recently encountered an issue (in a
branch) where invalid blocks were being retrieved from a cache. A regex
check on any of the base64 encoded fields would have caught it. To
address that, this commit adds regex check on the block 'indep_hash' and
'previous_block' fields.