docs(diagrams): add Makefile to render PlantUML and Graphviz PE-4997
We'd like to use PlantUML and in particular C4 to illustrate gateway
architecture so that it's easier for new developers to get up to speed
on it. This adds a Makefile that downloads the PlantUML jar and uses it
to render PlantUML diagrams as PNGs. Also, since I copied this from
another project that had the functionality already and it's generally
useful, it renders Graphviz dot diagrams too.