fix(range): return 416 instead of 200 for multiple range requests PE-4908
Since we're not fully supporting multiple range requets yet, we were
initially returning 200 along with full data for them requests. However,
it was pointed out that despite the 200 status code, clients may not
handle this correctly. This is likely the case, so this commit switches
the 200 to 416 (without data) for clarity.

This commit also switches some log messages for invalid ranges from
errors to warnings. We're trying to reserve error messages for potential
issues (bugs or ops problems). Invalid user input doesn't qualify.