feat(docker): restart containers when they fail PE-5129
Ideally we never need to restart containers, but crashes can happen. In
particular, if you fail to configure you memory settings correctly and
spin up a lot of worker processes it can kill the node process. This
change guards against that by restarting the containers when they exit
with an error code. We use a max of 5 consequetive restart to prevent
restart loops since those usually indicate an issue that can not be
resolved by restarts alone.