fix(graphql sqlite): correctly paginate TXs with duplicate tags PE-5140
If we get duplicate tags (same name + value combo) on a single TX or
data item, the joins to tags tables cause multiple rows to be returned
for the same TX. Those multiple rows are later turned into a single row
by a UNION. Since we use limits based on page size in the inner queries
this results in fewer rows than the page size being returned in the
outer query. That means when the count of rows is compared to the page
size to determine if we should indicate there's another page it returns
false. The fix for this is to use SELECT DISTINCT in the inner queries
to ensure they always return unique rows.