fix(ans-104): ensure that ans-104 streams are destroyed on error PE-5088
With ans-104 unbundling enabled, we've occasionally observed a large
increase in disk usage that disappears after the gateway service is
stopped. Our best guess is that this is related to hanging on to bundle
temporary file handles. This change attempts to address that by ensuring
temporary file streams are always destroyed when errors occur.