feat(docker): specify image tags in .env PE-5040
We want to start having end users use pre-built images for consistency
and reproduceablility. The most straightforward way to specify image
tags without needing to edit docker-compose.yaml on every release is to
pass them in as environment variables. docker-compose's service level
env_file support won't work for this since environment variables
specified there aren't available for interpolation in the
docker-compose.yaml file. If we want to avoid a docker-compose wrapper
script, we have to specify them in .env, so that's what we're doing.
Unfortunately this means users will need to move their .env to
.env.local. This should be farely obvious since git will complain when
they try to pull if they haven't done this, but we will have want to
alert users and document this thoroughly for the next release.