fix(env): move image tags out of .env to docker-compose.yaml PE-5198
Revert earlier change to put images tags in .env. This means user
specified configuration has been moved back to .env from .env.local.
This was done because .env.local environment variables are only
available within the services, but the user specified configuration
options are a mix of things that are passed to the services and used in
docker-compose.yaml. This means we'll need to update image tags in
docker-compose.yaml for each release, which is rather awkward, but still
workable. However, we'll also avoid the confusion of putting user
configuration in .env.local instead of .env.