fix(gateway-data-source): use 'Accept-Encoding: identity' for gateway data requests PE-5196
Axios changed it's default 'Accept-Encoding' behavior and it broke our
streaming cache. This restores the old behavior.

While somewhat annoying, this did expose the fact that we might not be
handling chunked responses properly since they won't have a
'Content-Length' header. Normally we won't receive those, but we might
in the future. We'll address that in a future commit.