feat(ans-104): disable unbundling workers if filter is '{ "never": true }'
The ANS-104 unbundling workers consume a significant amount of memory.
This change disables them entirely if the user's unbundle filter is set
to never match.

It also sets the default number of workers to 1 instead of 2 if
unbundling is enabled under the assumption that casual users will mostly
prefer lower memory consumption and power users will want to tune it
regardless.