perf(graphql sqlite): improve GraphQL performance predictability PE-5299
If allowed to make its own decisions, SQLite will sometimes pick poorly
performing join orders and indexes. This change forces the join order
using a CROSS JOIN and tells SQLite which indexes to use when executing
the joins using INDEXED BY. This ensures that we get the expected join
order based on tag order and ensures that a reasonable index is used for
the tag joins.

References:
- https://www.sqlite.org/optoverview.html
- https://www.sqlite.org/lang_indexedby.html