feat(chain sqlite): index stable transaction data offsets PE-2482
This change adds an 'offset' column to the stable_transactions table and
two worker processes - one that queries for missing offsets in the DB
and another that fetches them from Arweave nodes and indexes them in the
DB. Queries for missing offsets are made efficient through the use of a
partial index on the 'offset' where 'format' is 2 and 'data_size' is
greater than 0 since those are the only transactions that should have
offsets. To maintain a high indexing rate when backfilling, the indexing
worker fetchs up to 10 offsets in parallel. It also skips reprocessing
duplicate items in the queue and maintains a maximum queue depth of 1000
to avoid having an unbounded number of promises in flight. For
simplicity, we currently only index offsets for stable transactions
given they represent the bulk of the data.

In the future this transaction offset index can be used to serve chunks
requested by absolute offset directly from contiguous data sources
instead of making requests to Arweave nodes.