fix(arweave-client): drop poa2 from blocks to reduce cache size PE-5551
The block poa2 field was added in Arweave 2.6 and contains a 256KB chunk
when it is present and used (poa2 is not used on every block). Since
gateways do not index any data from it or validate blocks themselves
(they rely on a trusted Arweave node) we should drop it before caching
to conserve memory and disk space.