chore(redis): default to not saving to disk PE-5602
Users on memory constrained devices are occassionally experience issues
with Redis running out of memory when loading data from disk. Since we
only use Redis as an ephemeral cache, it's okay for us to not persist
cached data to disk. This change modifies the ar-io-node default Redis
config to not persist to disk. Advanced users can change this behavior
using the EXTRA_REDIS_FLAGS environment variable.