fix(arns): skip arns resolution when subdomains <= root subdomains PE-5749
This can happen when multiple root subdomains are present and the
healthcheck is called for localhost. In that case, attempting ArNS
resolution will throw an error when performing a string length check on
a undefined subdomain past the end of subdomains array. The fix is to
only perform ArNS resolution when there are more subdomains than are
present in the root domain.

Fixes #88