feat(data handlers): stop processing data stream for HEAD requests
When doing a HEAD request for `/tx-id` or `/raw/tx-id` endpoints the
server will not process the data stream while responding.
It will use the cached data information to set the headers and if no
cache is found it will process the data from the referenced gateway and
cache it.