feat(chunks): don't throw exceptions in broadcastChunks PE-5692
This alters the error handling behavior in broadcastChunks to avoid
throwing exceptions when the chunk broadcast success threshold has not
been met. This allows us to return more information to the caller in the
return value (still subject to change, we could also include this in a
thrown value). This change also adds a chunkBroadcaster interface (name
subject to change) so we can more easily support other chunk broadcast
mechanisms in the future.