feat(graphql clickhouse): add ClickHouse DB adapter PE-6488
This change adds support for a new ClickHouse based DB backend. It can
be used in combination with the SQLite DB backend to enable batch
loading of historical data from Parquet. It also opens up the
possibility of higher DB performance and scalability. In its current
state it should be considered a technology preview. It won't be useful
to most users until we either provide Parquet files to load into it or
automate flushing of the SQLite DB to it (both are planned in future
release). It also is not intended to be standalone solution. It supports
bulk loading and efficient GraphQL querying of transactions and data
items, but it relies on SQLite (or potentially another OLTP in the
future) to index recent data. These limitations allow greatly simplified
schema and query construction. Querying the new ClickHouse DB for
transaction and data items via GraphQL is enabled by setting the
'CLICKHOUSE_URL' environment variable.